inject: parallelize injection loading

+++ src/trx/thread_pool.h
@@ -1,13 +1,13 @@
 #pragma once
 
-#include <stddef.h>
+#include <stdint.h>
 
 typedef struct THREAD_POOL THREAD_POOL;
 typedef void (*THREAD_FUNC)(void *userdata);
 
 // Create a thread pool with the given number of worker threads.
 // Returns nullptr on failure.
-THREAD_POOL *ThreadPool_Create(size_t num_threads);
+THREAD_POOL *ThreadPool_Create(int32_t num_threads);
 
 // Destroy a thread pool, freeing all resources.
 // Blocks until all worker threads have exited.
